+++ pic24-dspic33-freertos-demo/Demo/dsPIC33C_SLAVE_MPLAB/slave/SlaveImage/RTOSDemo_Slave.h
@@ -1,0 +1,1 @@
+extern unsigned char RTOSDemo_Slave[] __attribute__((space(psv)));