+++ changelog.md
@@ -1,6 +1,7 @@
-# Project Name or repo name release version
+# PIC24 DSPIC33 FREERTOS DEMO v1.0.0
 ### Release Highlights
 
+Release of the PIC24 DSPIC33 FreeRTOS  Demo.
 
 
 ### Features Added\Updated